+++ package.json
@@ -26,6 +26,8 @@
         "check": "npm run lint && npm run svelte-check",
         "unittest": "deno test -A --no-check --coverage=cov_profile  --v8-flags=--expose-gc  --trace-leaks  ./src/",
         "test": "vitest run",
+        "test:unit": "vitest run --config vitest.config.unit.ts",
+        "test:unit:coverage": "vitest run --config vitest.config.unit.ts --coverage",
         "test:install-playwright": "npx playwright install chromium",
         "test:install-dependencies": "npm run test:install-playwright",
         "test:coverage": "vitest run --coverage",
